+++ Android/app/src/main/res/layout/activity_perfil_usuari_perfil_editar.xml
@@ -4,81 +4,60 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PerfilUsuariPerfilEditar">
+    tools:context=".PerfilUsuariPerfilEditarActivity">
 
     <Button
         android:id="@+id/bt_ActivarOpcionsPerfil"
         android:layout_width="143dp"
         android:layout_height="41dp"
-        android:layout_marginStart="220dp"
-        android:layout_marginLeft="220dp"
-        android:layout_marginBottom="400dp"
+        android:layout_marginTop="32dp"
         android:text="@string/editProfile"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.396" />
+        app:layout_constraintEnd_toEndOf="@+id/imageView2"
+        app:layout_constraintStart_toStartOf="@+id/imageView2"
+        app:layout_constraintTop_toBottomOf="@+id/imageView2" />
 
     <ImageView
         android:id="@+id/imageView2"
         android:layout_width="108dp"
         android:layout_height="90dp"
+        android:layout_marginTop="32dp"
         android:contentDescription="@string/todo"
-        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.93"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.025"
         app:srcCompat="@drawable/dibuixfotoperfil" />
 
-    <TextView
-        android:id="@+id/textSector"
-        android:layout_width="105dp"
-        android:layout_height="48dp"
-        android:text="@string/textview"
+    <Spinner
+        android:id="@+id/spinnerBotigues"
+        android:layout_width="302dp"
+        android:layout_height="68dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.107"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.373" />
+        />
 
-    <TextView
-        android:id="@+id/textView5"
-        android:layout_width="105dp"
-        android:layout_height="44dp"
-        android:text="@string/textview"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.108"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textSector"
-        app:layout_constraintVertical_bias="0.063" />
 
-    <CheckBox
-        android:id="@+id/checkBoxConsumidor"
+    <LinearLayout
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/consumidor"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/checkBoxBotiguer"
-        app:layout_constraintHorizontal_bias="0.41"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.084" />
+        android:layout_height="0dp"
+        android:layout_marginTop="32dp"
+        android:orientation="horizontal"
+        app:layout_constraintEnd_toEndOf="@+id/bt_ActivarOpcionsPerfil"
+        app:layout_constraintStart_toStartOf="@+id/bt_ActivarOpcionsPerfil"
+        app:layout_constraintTop_toBottomOf="@+id/bt_ActivarOpcionsPerfil">
 
-    <CheckBox
-        android:id="@+id/checkBoxBotiguer"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/botiga_negoci"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.477"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.084" />
+        <CheckBox
+            android:id="@+id/checkBoxConsumidor"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/consumidor" />
+
+        <CheckBox
+            android:id="@+id/checkBoxBotiguer"
+            android:layout_width="147dp"
+            android:layout_height="wrap_content"
+            android:text="@string/botiga_negoci" />
+    </LinearLayout>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
